+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- NAMD-GPU recipe
 
 ## [1.0.0] - 2016-09-22
 ### Added